Update snirf_specification.md
Added definition of the H5T.C_S1 as the String datatype.  This is what Matlab uses.  I figured out how to do this in my new C# code.  I am ok with still supporting CHAR arrays as "strings" but Jay and I found that HDF5 supports this C definition of a string (null terminated char array) as a predefined complex data type.

I also redefined the MetaDataTags as
/nirs/MetaDataTags(j).key and /nirs/MetaDataTags(j).value
I know Qianqian wants a string array but Jay and I could not figure out how to store a 2D array of strings in HDF5.  This is the only way we know how to currently do it from a programming stance.  Also, the use of a seperate value field means it does not need to be a string and so Key must be a string, but value can be string or numeric.